backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.2
Werkzeug==3.0.1
langchain==0.1.12
langchain-openai==0.0.8
openai==1.12.0
perplexity-python==0.1.0 